+++ firestore_api_dart/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Support snapshot metadata storage for later usage
+* Support for startAt, startAfter, endAt, endBefore collection queries
+
 ## 1.0.0
 
 * Initial Release